Initiate game from start button
Instead of having the game started automatically, add a start button to
initiate the game. When the game starts, clear the existing grid items
before repopulating with the new orders. When the game has ended because
the countdown timer has expired or the user has clicked the ready
button, stop processing grid presses.